Did 4-byte alignment in rt_mem_create()
v1.6.5
