+++ rtos/rt_memory.h
@@ -10,5 +10,6 @@
 
 #include "rt_TypeDef.h"
 
+void rt_mem_create(P_MEM pool, char *memory, unsigned int size);
 
 #endif /* RT_MEMORY_H_ */
